\name{arcAngle}
\alias{arcAngle}
\title{The counterclockwise arc of two points relative to a third}
\description{Compute the beginning and ending angles for an arc in radians relative to a third point}
\usage{arcAngle(x, y, center)}
\arguments{
  \item{x}{the xy coordinates of the first point}
  \item{y}{the xy coordinates of the second point that is counterclockwise from the first}
  \item{center}{the xy coordinates of the reference point}
}
\value{a two-item numeric vector containing the angle in radians of the two points relative to the reference point}
\author{Elliot Noma}
